Remove declineSupport() in server code.
Add isRegistered() to API

The client now asks the server "Am I registered?".  This avoids having
to be root to touch the file that marks this machine as registered.  Its
another network call, but allows blessings to work as a normal user.  I do
not believe this opens any new security issues as the server's DB
is always authoritative for answering the question about is this machine
registered.  You can't ask for specific machines either, only for the
client sending the HTTP GET request.


git-svn-id: https://svn.linux.ncsu.edu/svn/cls/trunk/xmlrpc@1606 35bbe6b0-b001-0410-9f0e-dd3f93706181