I've enjouyed working with cherrypy.  However, it of course just f'ing sucks when you believe that web apps should be run via apache.
This works with cherrypy's builtin web server and almost works with the
modpython WSGI gateway.  However, the cherrypy folks and the modpython
folks do not understand WSGI in the same way.  So, I get to deal with
cherrypy not handling the SCRIPT_NAME option correctly.  Right now
cherrypy.request.base isn't giving me the proper results which makes
the WGSI run app not be navigable.  


git-svn-id: https://svn.linux.ncsu.edu/svn/cls/trunk/ncsu-rlmtools@2055 35bbe6b0-b001-0410-9f0e-dd3f93706181