Figured this would bite evntually.  It is possible to have different key representations that are the same key.
I need to create a function to compare keys the right way.  These are RSA
keys so we need to compare N and E of two public keys to see if they match.

For the time being we compare the key stored in the database to the
reformated (key.exportKey()) and the key that the client gives us directly.
At this point in the code we have verified that the key the client sent
is a legit key and that the signature is valid.  When those two facts
are true we know that the client is itself legit.  All we have to do is
find the matching key in the database to locate the client's information.


git-svn-id: https://svn.linux.ncsu.edu/svn/cls/trunk/ncsu-rlmtools@2073 35bbe6b0-b001-0410-9f0e-dd3f93706181