client.py: Don't call the isRegistered() API method needlessly as it is expensive in the worst case.
server.py: We need to store the key text the client sends us so that we can
use the key text as a unique identifier.  The key text may change with 
different pickle format versions.  (Ie, python 2.2)  As long as we 
validate the key is a liget key text and check the signature we are
safe comparing and inserting (on register) that text into the db.



git-svn-id: https://svn.linux.ncsu.edu/svn/cls/trunk/ncsu-rlmtools@2077 35bbe6b0-b001-0410-9f0e-dd3f93706181