Add a getDeptBcfg2 API call
Clients and Web-Kickstart need to figure out how to init the Bcfg2
configuration manager and set it up with the right department.  So
any unauthorized client may call this API method to get that information
for a specific department.  The init string should be a %(foo)s-style
python template string WITHOUT any passwords to auth to the Bcfg2
repository.  Those are to be held and provided elsewhere.