RLMTools will now always follow host department changes.
Departments are controled by what's in the RLMTools attributes for that
host.  There is no real connection with the dept as referenced in the
realmlinux table.  Bcfg2 believes the attribute version and lays that down
on the client.  Now the client's will check in with their dept name
which will update that host's row in the realmlinux table.  Therefore,
the host will always appear in the dept on the webapp where you set
the dept attribute to be.