Add commit() calls to the DB dict wrappers for Bcfg2
We are doing a ton of selects and even selects cause locks.  Once
we iterate through an entire table a few times it looks like we can
start having database issues even on read-only selects.  Call commit()
after each query to the dict wrappers that result in a database hit.