ncsuclient: Improve message queuing to RLMTools
RLMTools removes clients that it hasn't heard from in 30 days.  We
remove old messages or corrupt messages at 30 days.  So as soon as a bad
message has prevented communication with RLMTools has finally aged
out...so our host has been deleted in RLMTools.  We check back in and
are set as no-support.  The timeout for bad messages is no 20 days.

We also add in a check for the SHA1 hash of a 0 length message file.
This is a common form of corruption so lets just detect it and unlink
the file.