Rebuild authentication system.
Authentication is still based on WRAP but we now parse all of it
ourselves.  This will allow us to run under any web server we chose, or
even the Flask test web server with WRAP authentication.

There is now one global Auth() object for all to use.  flask.g.auth