Rework authN/authZ code for simplicity
To force the user be to autheNticated just call

  g.auth.require()

That's the Auth class for the specific user/request.  To check
that the user is authoriZed for a specific department level
you can call any of:

  isREADby(dept)
  isWRITEby(dept)
  isADMINby(dept)

These functions also call g.auth.require() and force the authentication
step.  These functions do not return if the user is not authN'd/authZ'd.
They call Flask's abort() method and return an HTTP error.